+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Start Docker container if exists, but not running ([#243](https://github.com/arcavios/scooze/pull/243))
 - Add basic logging to some of our modules ([#249](https://github.com/arcavios/scooze/pull/249))
+- Add count_pips for Deck and DeckPart ([#251](https://github.com/arcavios/scooze/pull/251))
 
 ### Changed